Major changes
- Added ConstantBuffers:
     * glUniformBuffers for glsl shaders
          uniform ViewProjectionVSUniform
     * dxConstantBuffers for hlsl shaders
          cbuffer ViewVSConstant
          cbuffer ProjectionVSConstant
- Added Orthographic camera
- Added math library (glm)

- Fixes:
     * Renderer::Start
     * Window::SetDisplayMode
     * glShader destructor
     * glGraphicsContext constructor
     * dxGraphicsContext destructor

- Changes:
     * Renderer::DrawQuad
     * Window class: getters/setters are now static
     * dxGraphicsContext: members are now non-static
     * Changes to Demo & Sandbox projects
     * Renamed some functions and variables

- Removes:
     * WindowState
     * ExitCodes ( TerminationReq return value is now -1 ) I will add different return
          values based on the type of exception, eg. ImGui, DX/GL calls

+++ Light Engine/src/Engine/Layers/LayerStack.h
@@ -23,7 +23,7 @@ namespace Light {
 		static void AttachLayer(Layer* layer);
 		static void DetatchLayer(Layer* layer);
 
-		static unsigned int GetSize() { return s_Layers.size(); }
+		static unsigned int GetSize() { return static_cast<unsigned int>(s_Layers.size()); }
 
 		std::vector<Layer*>::iterator begin() { return s_Layers.begin(); }
 		std::vector<Layer*>::iterator end() { return s_Layers.end(); }
